feat: 퀘스트 클리어 경험치 골드 Long 타입으로 변경
RedisTemplate 사용 방식 변경

+++ src/main/java/todayquest/user/entity/UserInfo.kt
@@ -59,7 +59,7 @@ class UserInfo(
         coreTime = LocalTime.of(dto.coreTime, 0, 0)
     }
 
-    fun updateExpAndGold(questType: QuestType, earnedExp: Int, earnedGold: Int) {
+    fun updateExpAndGold(questType: QuestType, earnedExp: Long, earnedGold: Long) {
 
         val multiplier = when (questType) {
             QuestType.MAIN -> 2
